Define an interface method #refresh_content.
This method take the place of populate in master as well
as refresh_generated_content in this branch. The idea is
that all accounts are added because they generate content
for the system by some mechanism specfic to the account type.

By renaming the method we establish what it means to "populate"
an account. This then trickles up in order to make our channels
much cleaner and less concerned with the underlying account type.