Update requirements and tweak process_mentions.
Process mention now takes a polymorphic destination in addition to a channel id.
In the future we will need to tweak this and abstract the logic into a library class. The
Queue itself will then only be responsible for processing the job in the background and
selecting the polymorphic target.