Add proposed style guide to repo.
Hey guys, this is an early draft of a style guide primarily for Ruby in
which I've essayed a lot of techniques and approaches that will make our
codebase better and more agile.

This guide contains more than just rules for code formatting, it also
has suggestions for how to structure and think about program design and
structure as well as testing practices.

The business value these conventions will bring is not difficult to
measure because the rabid pace of development that they enable will make
back the initial effort of habitualizing them tenfold. The floor is open
for suggestions so if you have things you wish to add or points you
don't agreee with please put them forward. I would like to have an
agreed upon and formal style guide by the first release, if not by
public beta.