Extract mention processing logic into class.
Mention processing is now handled by the Mention::Processor class. An
instance of this class takes a mention and a destination where the
destination is any class that implements the push_recent instance
method like the trending channel and twitter account classes do.
Once this class is instantiated calling perform will run the processor.
If it succeeds nothing more need be done, you can check if it failed
with the #failed? instance method. This will return true if there were
errors during processing, otherwise false. If there were errors you can
get a text representation of them with the #errors instance method.