Add method for running twitter client as user.
Aji::Account::Twitter#authorize_with_twitter!

This method takes an optional block. If the block is given it will run
with the global twitter client as the authorized user, then reset back
to being authorized only with our app credentials. If no block is given
then this method MODIFIES GLOBAL STATE so that this is the active
twitter user. Use this with care!