Workaround bug in Hashie by using #key?
There is a bug in Hashie::Mash (now used by Grape for params) which
causes `params.has_key? :symbol` to incorrectly return false. We work
around this for now by using the #key? method which works, I have opened
a pull request so that mashes behave as expected and we can revert to
has_key? once it lands.