Isolate Category specs.
I'm starting to feel as though the featured specs in particular, and
many of the shared specs are approaching acceptance/integration level
status as they test the end result rather than the behavior of the
process. I would like to see them move out of the unit tests and into
the integration tests as we design better software where the tests need
less repeating (i.e. test the module in abstract during unit tests, and
invoke the shared example during integration/acceptance.)