Remove spec test for invalid behavior.
Queues::RefreshChannel will not work when passed the ID of an account.
The reason being that Resque will get a _Channel_ by id thus passing in
the ID of an account will not result in the desired behavior. A similar
spec will be added when an Account analogue to Queues::RefreshChannel is
written.