Fix bug in unauthorized Twitter accounts.
I accidentally neglected unauthorized users in my tests of the twitter
api. Which was pretty silly of me. I've added a token test for this for
now but in reality *all* of the twitter api methods should work
unauthorized save for #video_mentions_in_feed (and maybe even it for
some users??) unless a twitter user is protected. In order to accomplish
this I had to torture the constructor a little bit and take advantage of
some 1.9isms I'm not wholly proud of. I'm not sure if this is a band-aid
solution or the wave of the future for me. We shall see.

[#525 state:resolved]