Remove Cucumber/Capybara and replace with Spinach.
A while ago I took a stab at setting up cucumber for integration
testing. It was convoluted and difficult and I never really took it
anywhere. Spinach uses the same language for the tests but is slightly
cleaner when it comes to the definitions. Since it is still a departure
from spec tests I'd like @tpun to take a look at it when I get a few
more written and compare the style and clarity with integration spec
tests. I feel like the Cucumber/Spinach ones are more readable.

[#607]