Refactor #create_if_existing to be resistant to race conditions.
Well Thomas, you won out in the end, I do the check on the uid before
instantiating the account object. The reason, we were creating the new
object just for the check and then discarding it to use find_or_create.
In good Object Oriented code, we'd just persist the new account if it
was a valid youtube account, since we have to perform the race-check
anyway, I just use the YoutubeAPI singleton object (which is there
specifically for stuff like this).