Modify API tracker to prioritize different methods.
APITracker now takes an extra optional hash method_limits which
restricts the named methods to a proportion of the total. We're using
this to scale back on our youtube pushing in order to increase get
throughput. I've set the value for youtube post requests at 0.4, which
means that after 400 of our 1000 available requests, post will throttle
but get will go through.